+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 1.3.14 (unreleased)
 -------------------
 
-- Connect Users & Groups control panel settings to their registry entries.
+- Remove no-longer-used properties from portal_properties
   [esteele]
 
 - Remove plone_forms skins folder for 5.0 rc1